requirements.txt
pytest
allure-pytest
requests
faker
